pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-insight"
version = "0.1.0"
description = "Image Intelligence Service prototype"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.36",
  "pydantic>=2.9.0",
  "python-multipart>=0.0.9",
  "pillow>=10.4.0",
  "requests>=2.32.0",
  "redis>=5.2.0",
  "rq>=1.16.2",
  "minio>=7.2.12",
  "psycopg2-binary>=2.9.10",
  "jinja2>=3.1.4",
  "weasyprint>=66.0",
  "prometheus-client>=0.20.0",
  "vcrpy>=6.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_configs = true
disable_error_code = [
  "attr-defined",
  "no-any-return",
  "arg-type",
  "dict-item",
  "return-value",
  "import-untyped",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=60"
